+++ receiver/hostmetricsreceiver/internal/scraper/pagingscraper/internal/metadata/generated_metrics_v2.go
@@ -13,7 +13,7 @@ type MetricSettings struct {
 	Enabled bool `mapstructure:"enabled"`
 }
 
-// MetricsSettings provides settings for paging metrics.
+// MetricsSettings provides settings for hostmetricsreceiver/paging metrics.
 type MetricsSettings struct {
 	SystemPagingFaults      MetricSettings `mapstructure:"system.paging.faults"`
 	SystemPagingOperations  MetricSettings `mapstructure:"system.paging.operations"`
@@ -322,6 +322,16 @@ func (mb *MetricsBuilder) Reset(options ...metricBuilderOption) {
 	for _, op := range options {
 		op(mb)
 	}
+}
+
+// NewMetricData creates new pdata.Metrics and sets the InstrumentationLibrary
+// name on the ResourceMetrics.
+func (mb *MetricsBuilder) NewMetricData() pdata.Metrics {
+	md := pdata.NewMetrics()
+	rm := md.ResourceMetrics().AppendEmpty()
+	ilm := rm.InstrumentationLibraryMetrics().AppendEmpty()
+	ilm.InstrumentationLibrary().SetName("otelcol/hostmetricsreceiver/paging")
+	return md
 }
 
 // Attributes contains the possible metric attributes that can be used.