+++ receiver/postgresqlreceiver/scraper.go
@@ -82,9 +82,7 @@ func (p *postgreSQLScraper) scrape(ctx context.Context) (pdata.Metrics, error) {
 	}
 
 	// metric initialization
-	md := pdata.NewMetrics()
-	ilm := md.ResourceMetrics().AppendEmpty().InstrumentationLibraryMetrics().AppendEmpty()
-	ilm.InstrumentationLibrary().SetName("otelcol/postgresql")
+	md := p.mb.NewMetricData()
 	now := pdata.NewTimestampFromTime(time.Now())
 
 	var errors scrapererror.ScrapeErrors
@@ -106,7 +104,7 @@ func (p *postgreSQLScraper) scrape(ctx context.Context) (pdata.Metrics, error) {
 		p.collectDatabaseTableMetrics(ctx, now, dbClient, errors)
 	}
 
-	p.mb.Emit(ilm.Metrics())
+	p.mb.Emit(md.ResourceMetrics().At(0).InstrumentationLibraryMetrics().At(0).Metrics())
 	return md, errors.Combine()
 }